// web/src/app/model/form-type/add-form-type-model.ts
/* Auto Generated Code By AutoCodeGen Jabil © 2019 */
import { BaseModel } from '../base/base-model';
import { Constants } from '../../shared/constant/global';
import { FormType } from './form-type';
export class AddFormTypeModel extends BaseModel {
    formTypeId: number;
    type: string;
constructor(formType?: FormType) {
    super(formType);
    if (formType) {
        this.formTypeId = formType.formTypeId;
        this.type = formType.type;
    } else {
        this.formTypeId = 0;
        this.type = Constants.Empty;
    }
  }
}